[Bitmap] Title bar and Windows control buttons
Applied the svg files to the theme. I have decreased size of "x" mark in close button to match better a Luna style, and removed the gradient from title bar (it doesn't look good when applied on theme), also changed a Start button a bit to match better with the new taskbar.